+++ devicetypes/smartthings/smartsense-open-closed-sensor.groovy
@@ -198,7 +198,7 @@ private Map parseIasMessage(String description) {
 }
  
 def getTemperature(value) {
-	def celsius = Integer.parseInt(value, 16) / 100
+	def celsius = Integer.parseInt(value, 16).shortValue() / 100
 	if(getTemperatureScale() == "C"){
 		return celsius
 	} else {
